+++ hutool-captcha/src/main/java/cn/hutool/captcha/AbstractCaptcha.java
@@ -60,7 +60,7 @@ public abstract class AbstractCaptcha implements ICaptcha {
 	/**
 	 * 背景色
 	 */
-	protected Color background;
+	protected Color background = Color.WHITE;
 	/**
 	 * 文字透明度
 	 */