+++ hutool-captcha/src/main/java/cn/hutool/captcha/CircleCaptcha.java
@@ -82,8 +82,8 @@ public class CircleCaptcha extends AbstractCaptcha {
 
 	@Override
 	public Image createImage(String code) {
-		final BufferedImage image = new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB);
-		final Graphics2D g = ImgUtil.createGraphics(image, ObjectUtil.defaultIfNull(this.background, Color.WHITE));
+		final BufferedImage image = new BufferedImage(width, height, (null == this.background) ? BufferedImage.TYPE_4BYTE_ABGR : BufferedImage.TYPE_INT_RGB);
+		final Graphics2D g = ImgUtil.createGraphics(image, this.background);
 
 		try {
 			// 随机画干扰圈圈