+++ hutool-captcha/src/main/java/cn/hutool/captcha/GifCaptcha.java
@@ -3,6 +3,7 @@ package cn.hutool.captcha;
 
 import cn.hutool.captcha.generator.CodeGenerator;
 import cn.hutool.captcha.generator.RandomGenerator;
+import cn.hutool.core.img.ImgUtil;
 import cn.hutool.core.img.gif.AnimatedGifEncoder;
 import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.core.util.RandomUtil;
@@ -181,13 +182,11 @@ public class GifCaptcha extends AbstractCaptcha {
 	 * @return BufferedImage
 	 */
 	private BufferedImage graphicsImage(char[] chars, Color[] fontColor, char[] words, int flag) {
-		BufferedImage image = new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB);
+		final BufferedImage image = new BufferedImage(width, height, (null == this.background) ? BufferedImage.TYPE_4BYTE_ABGR : BufferedImage.TYPE_INT_RGB);
 		//或得图形上下文
-		Graphics2D g2d = image.createGraphics();
+		final Graphics2D g2d = ImgUtil.createGraphics(image, this.background);
 		try {
 			//利用指定颜色填充背景
-			g2d.setColor(ObjectUtil.defaultIfNull(this.background, Color.WHITE));
-			g2d.fillRect(0, 0, width, height);
 			AlphaComposite ac;
 			// 字符的y坐标
 			float y = (height >> 1) + (font.getSize() >> 1);