+++ hutool-captcha/src/main/java/cn/hutool/captcha/LineCaptcha.java
@@ -74,8 +74,8 @@ public class LineCaptcha extends AbstractCaptcha {
 	@Override
 	public Image createImage(String code) {
 		// 图像buffer
-		final BufferedImage image = new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB);
-		final Graphics2D g = GraphicsUtil.createGraphics(image, ObjectUtil.defaultIfNull(this.background, Color.WHITE));
+		final BufferedImage image = new BufferedImage(width, height, (null == this.background) ? BufferedImage.TYPE_4BYTE_ABGR : BufferedImage.TYPE_INT_RGB);
+		final Graphics2D g = ImgUtil.createGraphics(image, this.background);
 
 		try {
 			// 干扰线