+++ hutool-captcha/src/main/java/cn/hutool/captcha/ShearCaptcha.java
@@ -84,8 +84,8 @@ public class ShearCaptcha extends AbstractCaptcha {
 
 	@Override
 	public Image createImage(String code) {
-		final BufferedImage image = new BufferedImage(this.width, this.height, BufferedImage.TYPE_INT_RGB);
-		final Graphics2D g = GraphicsUtil.createGraphics(image, ObjectUtil.defaultIfNull(this.background, Color.WHITE));
+		final BufferedImage image = new BufferedImage(width, height, (null == this.background) ? BufferedImage.TYPE_4BYTE_ABGR : BufferedImage.TYPE_INT_RGB);
+		final Graphics2D g = ImgUtil.createGraphics(image, this.background);
 
 		try{
 			// 画字符串